Added unit tests for home_projects_page
